feat(backend) user and dtos ajust

+++ frontend/cue_swap/lib/ui/layouts/auth/views/widgets/student_form_data.dart
@@ -46,7 +46,7 @@ class StudentFormData extends StatelessWidget {
                 const SizedBox(height: 25),
                 TextFormField(
                       onChanged: (value){
-                        signUpFormProvider.studentData!.semester=value as int;
+                        signUpFormProvider.studentData!.semester=value;
                         signUpFormProvider.updateButtonState();
                       },
                       keyboardType: TextInputType.number,
